requirements.txt
-i https://pypi.org/simple
blinker==1.9.0 ; python_version >= '3.9'
cachecontrol==0.14.2 ; python_version >= '3.8'
cachetools==5.5.2 ; python_version >= '3.7'
certifi==2025.1.31 ; python_version >= '3.6'
cffi==1.17.1 ; python_version >= '3.8'
charset-normalizer==3.4.1 ; python_version >= '3.7'
click==8.1.8 ; python_version >= '3.7'
cloudevents==1.11.0
cryptography==44.0.2 ; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
deprecation==2.1.0
firebase-admin==6.7.0 ; python_version >= '3.7'
flask==3.1.0 ; python_version >= '3.9'
functions-framework==3.8.2 ; python_version >= '3.5' and python_version < '4'
google-api-core[grpc]==2.24.2 ; python_version >= '3.7'
google-api-python-client==2.167.0 ; python_version >= '3.7'
google-auth==2.39.0 ; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3 ; python_version >= '3.7'
google-cloud-firestore==2.20.2 ; python_version >= '3.7'
google-cloud-storage==3.1.0 ; python_version >= '3.7'
google-crc32c==1.7.1 ; python_version >= '3.9'
google-resumable-media==2.7.2 ; python_version >= '3.7'
googleapis-common-protos==1.70.0 ; python_version >= '3.7'
grpcio==1.71.0 ; python_version >= '3.9'
grpcio-status==1.71.0 ; python_version >= '3.9'
gunicorn==23.0.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10 ; python_version >= '3.6'
importlib-metadata==8.6.1 ; python_version >= '3.9'
itsdangerous==2.2.0 ; python_version >= '3.8'
jinja2==3.1.6 ; python_version >= '3.7'
markupsafe==3.0.2 ; python_version >= '3.9'
mlb-statsapi==1.8.1
msgpack==1.1.0 ; python_version >= '3.8'
packaging==24.2 ; python_version >= '3.8'
proto-plus==1.26.1 ; python_version >= '3.7'
protobuf==5.29.4 ; python_version >= '3.8'
pyasn1==0.6.1 ; python_version >= '3.8'
pyasn1-modules==0.4.2 ; python_version >= '3.8'
pycparser==2.22 ; python_version >= '3.8'
pyjwt[crypto]==2.10.1 ; python_version >= '3.9'
pyparsing==3.2.3 ; python_version >= '3.9'
python-dotenv==1.0.1 ; python_version >= '3.8'
requests==2.32.3 ; python_version >= '3.8'
rsa==4.9.1 ; python_version >= '3.6' and python_version < '4'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.4.0 ; python_version >= '3.9'
watchdog==6.0.0 ; python_version >= '3.9'
werkzeug==3.1.3 ; python_version >= '3.9'
zipp==3.21.0 ; python_version >= '3.9'
